# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Whispick API — anonymous video recommendation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whisps:
    get:
      operationId: listWhisps
      tags: [whisps]
      summary: List all whisps sent by the current user
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of whisps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Whisp"
    post:
      operationId: createWhisp
      tags: [whisps]
      summary: Send a new whisp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhispInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Whisp"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /whisps/{id}:
    get:
      operationId: getWhisp
      tags: [whisps]
      summary: Get a single whisp with tracking timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whisp detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteWhisp
      tags: [whisps]
      summary: Delete a whisp
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /whisps/{id}/replies:
    get:
      operationId: listWhispReplies
      tags: [whisps]
      summary: Get anonymous replies for a whisp
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhispReply"
    post:
      operationId: createWhispReply
      tags: [whisps]
      summary: Send an anonymous reply (from recipient or sender follow-up)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhispReplyInput"
      responses:
        "201":
          description: Reply created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispReply"
  /whisps/{id}/reveal:
    post:
      operationId: requestReveal
      tags: [whisps]
      summary: Sender requests to reveal identity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reveal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Whisp"
    patch:
      operationId: respondReveal
      tags: [whisps]
      summary: Recipient responds to reveal request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevealResponse"
      responses:
        "200":
          description: Reveal response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealResult"
        "400":
          description: No reveal was requested for this whisp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/whisps:
    get:
      operationId: adminListWhisps
      tags: [admin]
      summary: List all whisps across all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: deliveryMethod
          schema:
            type: string
          required: false
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: pageSize
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Whisps page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWhispListResponse"
  /admin/whisps/{id}:
    get:
      operationId: adminGetWhisp
      tags: [admin]
      summary: Get full detail on any whisp (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whisp detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWhispDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: adminDeleteWhisp
      tags: [admin]
      summary: Remove a whisp (moderation) (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
    Whisp:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        videoUrl:
          type: string
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoStartSeconds:
          type: ["integer", "null"]
        videoPlatform:
          type: ["string", "null"]
        videoTranscript:
          type: ["string", "null"]
        uploadedVideoId:
          type: ["string", "null"]
        deliveryMethod:
          type: string
        whisperChannel:
          type: ["string", "null"]
        circleId:
          type: ["string", "null"]
        recipientEmail:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        anonymousNote:
          type: ["string", "null"]
        senderAlias:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        status:
          type: string
        publicToken:
          type: string
        scheduledAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        watchedAt:
          type: ["string", "null"]
        revealRequested:
          type: boolean
        revealAccepted:
          type: ["boolean", "null"]
        appreciationResponse:
          type: ["string", "null"]
        appreciationRespondedAt:
          type: ["string", "null"]
        aiTakeaway:
          type: ["string", "null"]
        aiTakeawayStatus:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - senderId
        - videoUrl
        - deliveryMethod
        - status
        - publicToken
        - revealRequested
        - createdAt
    WhispInput:
      type: object
      properties:
        videoUrl:
          type: ["string", "null"]
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoStartSeconds:
          type: ["integer", "null"]
        videoPlatform:
          type: ["string", "null"]
        uploadedVideoId:
          type: ["string", "null"]
          description: An id from the sender's Media Library — an alternative to videoUrl. One of the two is required.
        deliveryMethod:
          type: string
        whisperChannel:
          type: ["string", "null"]
        circleId:
          type: ["string", "null"]
        recipientEmail:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        anonymousNote:
          type: ["string", "null"]
        senderAlias:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
      required:
        - deliveryMethod
    WhispDetail:
      type: object
      properties:
        whisp:
          $ref: "#/components/schemas/Whisp"
        trackingEvents:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/WhispReply"
      required:
        - whisp
        - trackingEvents
        - replies
...
    WhispReply:
      type: object
      properties:
        id:
          type: string
        whispId:
          type: string
        replyText:
          type: string
        fromRecipient:
          type: boolean
        videoUrl:
          type: ["string", "null"]
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoPlatform:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - whispId
        - replyText
        - fromRecipient
        - createdAt
    WhispReplyInput:
      type: object
      properties:
        replyText:
          type: string
        fromRecipient:
          type: boolean
      required:
        - replyText
    RevealResponse:
      type: object
      properties:
        accepted:
          type: boolean
      required:
        - accepted
    RevealResult:
      type: object
      properties:
        id:
          type: string
        revealRequested:
          type: boolean
        revealAccepted:
          type: boolean
      required:
        - id
        - revealRequested
        - revealAccepted
...
    TrackingEvent:
      type: object
      properties:
        id:
          type: string
        whispId:
          type: string
        eventType:
          type: string
        createdAt:
          type: string
      required:
        - id
        - whispId
        - eventType
        - createdAt
...
    WhispCategory:
      type: object
      properties:
        id:
          type: string
        whispId:
          type: string
        category:
          type: string
        rank:
          type: integer
        score:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - whispId
        - category
        - rank
        - score
        - createdAt
    AdminWhispListItem:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        senderEmail:
          type: ["string", "null"]
        videoUrl:
          type: string
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoPlatform:
          type: ["string", "null"]
        deliveryMethod:
          type: string
        whisperChannel:
          type: ["string", "null"]
        circleId:
          type: ["string", "null"]
        status:
          type: string
        scheduledAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        watchedAt:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        boostSpendUsd:
          type: ["string", "null"]
        categories:
          type: array
          items:
            $ref: "#/components/schemas/WhispCategory"
        createdAt:
          type: string
      required:
        - id
        - senderId
        - videoUrl
        - deliveryMethod
        - status
        - categories
        - createdAt
    AdminWhispListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminWhispListItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - items
        - total
        - page
        - pageSize
    AdminWhispDetail:
      type: object
      properties:
        whisp:
          $ref: "#/components/schemas/Whisp"
        senderId:
          type: ["string", "null"]
        senderEmail:
          type: ["string", "null"]
        senderFullName:
          type: ["string", "null"]
        trackingEvents:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/WhispReply"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/WhispCategory"
      required:
        - whisp
        - trackingEvents
        - replies
        - categories
...