# lib/api-zod/src/generated/types/whisp.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Whispick API — anonymous video recommendation platform
 * OpenAPI spec version: 0.1.0
 */

export interface Whisp {
  id: string;
  senderId: string;
  videoUrl: string;
  /** @nullable */
  videoTitle?: string | null;
  /** @nullable */
  videoThumbnail?: string | null;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /** @nullable */
  videoStartSeconds?: number | null;
  /** @nullable */
  videoPlatform?: string | null;
  /** @nullable */
  videoTranscript?: string | null;
  /** @nullable */
  uploadedVideoId?: string | null;
  deliveryMethod: string;
  /** @nullable */
  whisperChannel?: string | null;
  /** @nullable */
  circleId?: string | null;
  /** @nullable */
  recipientEmail?: string | null;
  /** @nullable */
  recipientPhone?: string | null;
  /** @nullable */
  anonymousNote?: string | null;
  /** @nullable */
  senderAlias?: string | null;
  /** @nullable */
  moodTag?: string | null;
  status: string;
  publicToken: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  openedAt?: string | null;
  /** @nullable */
  watchedAt?: string | null;
  revealRequested: boolean;
  /** @nullable */
  revealAccepted?: boolean | null;
  /** @nullable */
  appreciationResponse?: string | null;
  /** @nullable */
  appreciationRespondedAt?: string | null;
  /** @nullable */
  aiTakeaway?: string | null;
  /** @nullable */
  aiTakeawayStatus?: string | null;
  createdAt: string;
}