# lib/api-zod/src/generated/types/whispInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Whispick API — anonymous video recommendation platform
 * OpenAPI spec version: 0.1.0
 */

export interface WhispInput {
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  videoTitle?: string | null;
  /** @nullable */
  videoThumbnail?: string | null;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /** @nullable */
  videoStartSeconds?: number | null;
  /** @nullable */
  videoPlatform?: string | null;
  /**
     * An id from the sender's Media Library — an alternative to videoUrl. One of the two is required.
     * @nullable
     */
  uploadedVideoId?: string | null;
  deliveryMethod: string;
  /** @nullable */
  whisperChannel?: string | null;
  /** @nullable */
  circleId?: string | null;
  /** @nullable */
  recipientEmail?: string | null;
  /** @nullable */
  recipientPhone?: string | null;
  /** @nullable */
  anonymousNote?: string | null;
  /** @nullable */
  senderAlias?: string | null;
  /** @nullable */
  moodTag?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}